feat: add flutter adaptative layout
