+++ lib/ui/router/app_router.dart
@@ -6,7 +6,7 @@ import 'package:flutter_base/ui/features/auth/views/forgot_password/forgot_passw
 import 'package:flutter_base/ui/features/auth/views/login/login_page.dart';
 import 'package:flutter_base/ui/features/auth/views/sign_up/sign_up_page.dart';
 import 'package:flutter_base/ui/features/auth/views/verify_account/verify_account_page.dart';
-import 'package:flutter_base/ui/features/misc/components/scaffold_with_nav_bar.dart';
+import 'package:flutter_base/ui/features/misc/components/scaffold_with_navigation.dart';
 import 'package:flutter_base/ui/features/misc/views/main_page.dart';
 import 'package:flutter_base/ui/features/post/views/detail_post_page/detail_post_page.dart';
 import 'package:flutter_base/ui/features/post/views/posts/post_page.dart';
@@ -19,12 +19,12 @@ import 'package:flutter_base/ui/router/utils.dart';
 import 'package:go_router/go_router.dart';
 
 final _bottomBarItems = [
-  ScaffoldWithNavBarTabItem(
+  ScaffoldWithNavigationItem(
     icon: Icons.list,
     rootRoutePath: '/home',
     navigatorKey: GlobalKey<NavigatorState>(debugLabel: 'homeNav'),
   ),
-  ScaffoldWithNavBarTabItem(
+  ScaffoldWithNavigationItem(
     icon: Icons.account_circle,
     rootRoutePath: '/profile',
     navigatorKey: GlobalKey<NavigatorState>(debugLabel: 'profileNav'),
