+++ pubspec.yaml
@@ -26,6 +26,7 @@ dependencies:
   firebase_messaging: 14.3.0
   flutter:
     sdk: flutter
+  flutter_adaptive_scaffold: 0.1.3
   flutter_config:
     git:
       url: https://github.com/ByneappLLC/flutter_config.git
@@ -76,6 +77,7 @@ dev_dependencies:
   reactive_forms_generator: 1.0.0
 
 dependency_overrides:
+  archive: 3.3.2
   collection: 1.17.0
   intl: 0.17.0
